[Filestore]: support remove by parentId and name in filestore-client

+++ cloud/filestore/apps/client/lib/rm.cpp
@@ -16,6 +16,7 @@ class TRmCommand final
 private:
     TString Path;
     bool RemoveDir = false;
+    ui64 NodeId = 0;
 
 public:
     TRmCommand()
@@ -24,6 +25,11 @@ public:
             .Required()
             .RequiredArgument("PATH")
             .StoreResult(&Path);
+
+        Opts.AddLongOption("node")
+            .DefaultValue(0)
+            .RequiredArgument("ID")
+            .StoreResult(&NodeId);
 
         Opts.AddLongOption('r', "recursive")
             .NoArgument()
@@ -35,14 +41,20 @@ public:
         auto sessionGuard = CreateSession();
         auto& session = sessionGuard.AccessSession();
 
-        const auto resolved = ResolvePath(session, Path, false);
+        // If nodeId is set, treat it as a nodeId and path as a name. Otherwise,
+        // resolve path to get nodeId and name.
+        if (!NodeId) {
+            const auto resolved = ResolvePath(session, Path, false);
+            Y_ENSURE(resolved.size() >= 2, "can't rm root node");
 
-        Y_ENSURE(resolved.size() >= 2, "can't rm root node");
+            NodeId = resolved[resolved.size() - 2].Node.GetId();
+            Path = ToString(resolved.back().Name);
+        }
 
         auto request = CreateRequest<NProto::TUnlinkNodeRequest>();
 
-        request->SetNodeId(resolved[resolved.size() - 2].Node.GetId());
-        request->SetName(ToString(resolved.back().Name));
+        request->SetNodeId(NodeId);
+        request->SetName(Path);
         request->SetUnlinkDirectory(RemoveDir);
 
         auto response = WaitFor(session.UnlinkNode(
